Implement key rotation
This uses the preshared password to generate an initial key. That
preshared key is then used as a derivation key for key rotations.

This does not actually provide forward secrecy as there's a
deterministic relationship between the initial preshared key and
subsequent encryption keys. Instead, it is an iterative step towards
implementing an olm/signal ratchet forward secrecy mechanism because it
exercises dynamic key selection.