Livekit e2ee implementation
This is the current state of the livekit implementation. It's a work in
progress/open PR, but we're interested in moving forwards with it.
Therefore I've squashed all of the work into this commit so that when it
does get merged/accepted we can more easily replace this with the
finished work.